prototype/requirements_wsl.txt
numpy>=1.24
PyYAML>=6.0
matplotlib>=3.7
pandas>=2.0
jupyter>=1.0
